Changed from `double` to `float`

+++ lib/IMU/6-Axis-IMU.h
@@ -1,11 +1,8 @@
-
 #ifndef IMU_HEADER
 #define IMU_HEADER
 
-#include <Arduino.h>
-
 void imu_setup();
-void imu_record(double* accX, double* accY, double* accZ, double* gyrX, double* gyrY, double* gyrZ, double* tempIMU);
+void imu_record(float* accX, float* accY, float* accZ, float* gyrX, float* gyrY, float* gyrZ, float* tempIMU);
 void imu_print();
 
 #endif
